jira/requirements.txt
jira>=3.8
pyyaml>=6.0
python-dotenv>=1.0
